Install binaries, remove reference to /home/nemo
The binaries weren't being correctly installed, because the
%exclude of /usr/share/GetiPlay/bin was removing the binaries
even though they were explicitly added later. This was added
to remove a duplicate file warning which has now returned
(but doesn't seem to cause problems).

Also, removed all reference to /home/nemo and replaced them with
$HOME. They were littered all over the perl5 library install.
This takes one step closer to complying with the Harbour
guidelies:

https://harbour.jolla.com/faq