Improve refresh progress indicator
During refresh progress is now split into two stages, a download
stage and a processing stage. We don't know how much data either
stage will generate, but the download stage seems to be quite
variable, whereas we assume the processing stage will remain
relatively stable (because the number of programmes available
from one point in time to the next doesn't change too much).
Consequently, we use the download stage to fill the progress
meter partially, then fill the rest with the processing stage.
This hopefully gives a more consistent appearance for the user.