Remove DIR_PERLLOCAL "/bin" from get-iplayer path
This folder contains perl5 binary scripts, but none of these are
needed to run get-iplayer (thinks to @dinkypumpkin for pointing
this out.

TV/Radio refresh and download have all been tested with this.

However, it turns out the bin folder containing ffmpeg is
still needed (even though the full ffmpeg path is passed to
get-iplayer). get-iplayer fails to find ffmpeg without it.