Remove 'Switch' menu item, improved date handling
The 'Switch to TV' and 'Switch to Radio' menu items no longer
make sense given that both are availalbe on different tabs. The
menu item to swith between them has therefore been removed.

The date field in the programme info dialogue was only filled out
on downloading the metadata. However, the date added is retrieved
during a refresh, so it may make sense to use this as an
approximate date until the actual 'first broadcast' date has been
downloaded as metadata.