Pop ProgInfo dialogue on download
Downloading a programme automatically pops the info page off the
stack. This makes it much easier to swipe to the queue page, if
that's your inclination.

This also fixes an issue whereby the 'Download' button woudl be
active if the programme was queued but hadn't yet started
downloading. In this situation, there's no need to add it to
the queue again, so the button should be disabled.