Restructure log window for performance
Changing the height of the log text area, with the log text
aligned to the bottom of the label, caused terrible performance
when changing the height of the label (which happens when you
click on the log area, so that it can show the sharing options).

The fix has been to attach the label to the bottom of its parent
rectangle, and make this rectangle clip the contents. Now the
rectangle changes height (clipping the log text inside) but the
label containing the text stays the same height (it's just that
it's being clipped by its parent, so it looks like it's changing
vertical height). Miraculously, this fixes the performance issue.