Add perl-Digest-SHA as dependency
Without this, Mojolicious isn't picked up in the local lib, even
if Digest::SHA is installed in the local lib.

Unfortunately, it seems this has to be installed from the
package repositories, rather into local lib, for it to work.
There must be a way to properly bootstrap local lib without it, but
I couldn't figure it out.