Display audio/video play button only once downloaded
The button to play in an external app becomes available only once
a media file is fully downloaded. However, previously the button
to play using the built-in player was always available. This change
makes the button for the internal player appear at the same time
as the external player button.