Add 'All' option alongside national/regional/local
Previously the refresh type was restricted, so that only one of
national, regional or local programme types could selected at
once. The restriction was for performance reasons (the scrolling
suffered with so many items in th list).

This change adds an 'All' option, allowing all three types to be
refreshed and displayed at once. This is to test performance
again, given the programme has changed a fair bit since this
was last tested, in the hope the performance is now acceptable.