Ensure button icons are correct size across devices
The stock Sailfish icons have different sizes on different devices
with the following scaling factors:

1.0 (jolla phone), 1.25 (jolla c), 1.5 (tablet), 1.75 (xperia)

However, the application icons used were the same on all devices
(equivalent of 1.75 xperia scaling). The IconButton elements used
the sizes of the icons for the buttons (no explicit sizes were
given), resulting in different size icons on different devices,
and depending on which set (stock or app) they came from. The
icons all looked fine on the X because that's what I tested them
on.

This change sets the sizes of the IconButtons explicitly, so they
all appear the same size on all devices. Currently, the app images
are scaled down to suit other devices.