Obtain CoverAction image URL correctly
When instantiating the CoverAction, Sailfish doesn't seem to be
able to access the GetiPlay custom QQuickImageProvider, so the
custom image is missing from the 'replay' button.

This change shifts the decision over which directory to get the
images from (based on the pixel ratio of the device being used)
from the image provider and into the settings singleton. The
image provider queries the settings singleton to determine the
folder. As a result, the URL for the CoverAction image can be
extracted from the settings singleton, avoiding the need to query
the (inaccessible) image provider.