All index max connections to be set
When refreshing, Mojolicious generates failure warnings on Jolla One
and Jolla C if the number of simultaneous connections rises above
two. On the Xperia X it seems to be able to handle at least five.

This change makes the maximum number of connections during a refresh
user-configurable. Sensible settings of either 2 or 5 are set for the
devices that it's been tested on (Jolla One, Jolla C, Xperia X).