Fix dates, improve animation, fix localisation
When using PIDs, it's possible that no 'available' filed will be
returned in the programme info (because the media isn't available
yet). This is now captured and handled properly (the date is
shown as 'Unknown').

Decreasing the size of the column containing the pid entry field
removed the item, which also then removed the spacing between the
items, causing the animated line to jump upwards at the end. This
change sets the final size to 1 rather than 0, preventing this
from happening.

Some localisation issues have been fixed, specifically re-use of
string ids from the ProgInfo page that should have been fresh.