Allow PID entry section to fully dissappear
It turns out that the restructuring of the UI (removing the gaps
between items in the main column and moving them to the child
columns) means that allowing the PID entry to fully dissappear
(i.e. animating its height to zero) no longer causes the animation
to jump. Moving it to zero would make sense, since it may be
slightly more efficient.