Ensure trimmed log is trimmed to 50 lines
Previously the log file was trimmed to 50 lines each time a new
line was added to it. However, the check was skipped if the
log started with a newline, because the condition for whether a
newline had been found checked for indices greater than zero. It
should have checked for indices that included zero as well. This
has now been fixed.

In addition, the log is trimmed each time it's loaded (which only
happens when the app is started). This addresses the sitation
where the log file is changed outside the program.