Add get_iplayer, atomicparsley and perl libs to the build process
Previously get_iplayer, atomicparsley and the perl libs had to be built
separately and copied in to the project. This had many disadvantages,
not least that they weren't automatically updated to the latest
versions, and that the process needed to be performed separately for
each target the project was built for.

This change adds the dependencies to the build process, so they are
either built or imported for packaging as part of the build.

Also removes ffmpeg, which is now available in the official Sailfish
repositories, so doesn't need to be included in the package.